+++ src/main/java/org/nodens2k/tin/validation/CzechiaTinValidator.java
@@ -21,7 +21,7 @@ public final class CzechiaTinValidator extends AbstractCountryTinValidator {
   @Contract(value = "null -> false", pure = true)
   @Override
   public boolean isValid(String tin) {
-    tin = sanitise(tin);
+    tin = sanitise(tin, "CZ");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;
     }