+++ src/main/java/org/nodens2k/tin/validation/EstoniaTinValidator.java
@@ -20,7 +20,7 @@ public final class EstoniaTinValidator extends AbstractCountryTinValidator {
   @Contract(value = "null -> false", pure = true)
   @Override
   public boolean isValid(String tin) {
-    tin = sanitise(tin);
+    tin = sanitise(tin, "EE");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;
     }